% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eye.foodweb.cascade}
\alias{eye.foodweb.cascade}
\title{Builds a food web according to the cascade model}
\usage{
eye.foodweb.cascade(S = 100, C = 0.25)
}
\arguments{
\item{S}{The number of species (integer)}

\item{C}{The desired connectance (between 0 and 1)}
}
\value{
A connected (one piece) food web of connectance close to the desired value.
The food web is a list containing:
\describe{
\item{$links}{The trophic links from resource to consumer (L x 2 matrix)}
\item{$S}{The number of species}
\item{$L}{The number of links}
\item{$C}{The connectance L / (S choose 2)}
\item{$Model}{A code for the model used to generate the food web}
\item{$Cycles}{Boolean, whether cycles are present}
}
All the self-loops and double arrows (a eats b, b eats a) are removed. If the food web contains
no cycles, then it is sorted that all the coefficients would be in the upper-triangular part
of the corresponding adjacency matrix.

If the resulting network has connectance too different from the desired one, or it is not connected,
the function will call itself recursively. If too many recursions are performed, it will fail.
}
\description{
The cascade model is described in Cohen et al. "Community food webs: Data and theory", 1990.
The species are sorted in order from 1...S and each species has probability C of consuming any of the
preceding species.
}
\examples{
FW <- eye.foodweb.cascade() # S = 100, C = 0.25
FW <- eye.foodweb.cascade(30, 0.2)
FW <- eye.foodweb.cascade(S = 250, C = 0.314)
}

