% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eye.pairs.from.fourcorner}
\alias{eye.pairs.from.fourcorner}
\title{Sample pairs of interactions from the "Four-Corner" distribution described
in Allesina et al. 2014}
\usage{
eye.pairs.from.fourcorner(NumPairs = 10, mux = -1, muy = 0.5,
  sigmax = 1/4, sigmay = 1/4, rhoxy = -2/3)
}
\arguments{
\item{NumPairs}{The number of pairs to sample}

\item{mux}{The desired mean for the first marginal distribution}

\item{muy}{The desired mean for the second marginal distribution}

\item{sigmax}{The desired standard deviation for the first marginal distribution}

\item{sigmay}{The desired standard deviation for the second marginal distribution}

\item{rhoxy}{The desired correlation}
}
\value{
A NumPairs x 2 matrix of interaction strengths sampled from the distribution
}
\description{
Sample pairs of interactions from the "Four-Corner" distribution described
in Allesina et al. 2014
}
\examples{
mypairs <- eye.pairs.from.fourcorner() # default values
mypairs <- eye.pairs.from.fourcorner(NumPairs = 120)
}

