% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyeCutoff.R
\name{eyeCutoff}
\alias{eyeCutoff}
\title{Find values that exceed a given cutoff}
\usage{
eyeCutoff(eData, stdevs = NULL, col = NULL)
}
\arguments{
\item{eData}{A data frame}

\item{stdevs}{The cutoff to define extreme scores in units of standard deviation.}

\item{col}{Optional column. If provided, eyeCutoff will return the rows with missing values, for that column.}
}
\value{
Returns either a listing of columns with missing values (and totals) or one column with the missing rows.
}
\description{
Provide a data frame and a cutoff point (in units of standard deviation) and receive
a list of columns with extreme values, along with the number of values. If a column
is provided in the call, then this function will instead return a listing of rows in that
column that have missing values.
}
