% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyeDist.R
\name{eyeDist}
\alias{eyeDist}
\title{Eyeball the distributions}
\usage{
eyeDist(eData, brks = "Sturges")
}
\arguments{
\item{eData}{A data frame that includes at least one numeric or factored column.}

\item{brks}{Optional number of histogram breaks. Will default if empty.}
}
\value{
Returns a listing of columns for which 's' was pressed during viewing.
}
\description{
Accepts numeric and factored categorical variables and produces histograms or
bar plots that can be cycled through using '>' or '<'. Returns a listing of any
columns that are of interest. Columns are added to this list by pressing 's'.
}
