% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyeMissing.R
\name{eyeMissing}
\alias{eyeMissing}
\title{Find missing data}
\usage{
eyeMissing(eData, col = NULL)
}
\arguments{
\item{eData}{A data frame}

\item{col}{Optional column to be searched for missing values. Enter as either 'col_name' or type the column number.}
}
\value{
Returns either a listing of columns with missing values or missing values in a specified column.
}
\description{
If data frame is entered this function will return the columns with missing values
and the total number of missing values. If a column name is included in the call,
this function will instead return a list of rows that have missing values.
}
