% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyeNorm.R
\name{eyeNorm}
\alias{eyeNorm}
\title{Cycle through data to check distribution normality of numeric data}
\usage{
eyeNorm(eData)
}
\arguments{
\item{eData}{A data frame containing at least on numeric column.}
}
\value{
Returns a list of columns of interest, which is created by pressing 's'.
}
\description{
Provides qqplot, line and Shapiro-Wilks test. Cycle through data using '<'
or '>'. Add a column to a returned list by pressing 's'.
}
