% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyeScatter.R
\name{eyeScatter}
\alias{eyeScatter}
\title{Cycle through scatter plots}
\usage{
eyeScatter(eData, y)
}
\arguments{
\item{eData}{A data frame containing numeric and/or categorical variables}

\item{y}{The column to appear on the y-axis entered as either 'col_name' or simply by typing the column number.}
}
\value{
Returns a listing of columns for which 's' was pressed.
}
\description{
Accepts a data frame and one column to appear on the y-axis. Allows user to cycle
through columns, which will appear on the x-axis. This is done by pressing '<' or
'>'. User can save a list of columns of interest by pressing 's'. NOTE: It will make
categorical variables numeric whenever possible.
}
